requirements.txt
PyGithub==1.58.2
python-dotenv==1.0.0
rich==13.3.5
